+++ HISTORY.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+  * Improve string performance
   * Slightly improve speed for weak ETags over 1KB
 
 1.4.0 / 2014-09-21